major structure optimize
