+++ scrap_example.txt
@@ -1,0 +1,204 @@
+# --- --- --- Section Get Lesson List --- --- --- #
+# {
+#     "university_name": "",
+#     "term": 202401,
+#     "university_logo_pic": "",
+#     "name": "NAME",
+#     "type_count": [],
+#     "students_count": 7,
+#     "color_system": 3,
+#     "course": {
+#         "update_time": "",
+#         "name": "",
+#         "admin_id": 0,
+#         "university_id": 0,
+#         "type": 0,
+#         "id": 0
+#     },
+#     "teacher": {
+#         "user_id": 0,
+#         "name": "",
+#         "avatar": ""
+#     },
+#     "create_time": "",
+#     "university_id": 0,
+#     "time": "",
+#     "course_id": 0,
+#     "university_logo": "0",
+#     "university_mini_logo": "0",
+#     "id": 0,
+#     "is_pro": true,
+#     "color_code": 0
+# }
+
+
+# --- --- --- Section Download Lesson Video --- --- --- #
+# {
+#      "type": 14,
+#      "id": 7153416,
+#      "courseware_id": "909642544544463488",
+#      "title": "R8-三相-周期非正弦",
+#      "create_time": 1686274642000,
+#      "attend_status": true,
+#      "is_finished": true
+# }
+
+
+# --- --- --- Section Download Segment --- --- --- #
+# {
+#     "id": "743834725938342272",
+#     "code": "kszt_DdQU9sOod7o",
+#     "type": 2,
+#     "source": "th",
+#     "url": "https://kszt-playback.xuetangx.com/gifshow-xuetangx/73466bdb387702307504996781/f0.mp4?auth_key=1729778852-4128559473511008914-0-e0c959d1504f92ef5a5d45000f46330d",
+#     "start": 1666508813000,
+#     "end": 1666510612000,
+#     "duration": 1799000,
+#     "hiddenStatus": 0,
+#     "order": 0,
+#     "replayOssStatus": 0,
+#     "recordFileId": "",
+#     "recordType": "",
+#     "subtitlePath": ""
+# }
+
+
+# --- --- --- Section Download Lesson PPT --- --- --- #
+# {
+#     "code": 0,
+#     "msg": "OK",
+#     "data": {
+#         "lesson": {
+#             "id": "1267751345493205504",
+#             "title": "计算机网络原理  课堂提问（2）",
+#             "startTime": 1728964537223,
+#             "endTime": 1728965834350,
+#             "teacherIdentityId": "15753469",
+#             "classroom": {
+#                 "id": "3134428",
+#                 "name": "2024秋-计算机网络原理-2",
+#                 "pro": true
+#             },
+#             "course": {
+#                 "id": "1360043",
+#                 "name": "计算机网络原理"
+#             }
+#         },
+#         "fileSharing": {
+#             "count": 0,
+#             "cover": null
+#         },
+#         "teacher": {
+#             "identityId": "15753469",
+#             "avatar": "0",
+#             "name": "徐明伟",
+#             "number": "1998990267"
+#         },
+#         "replayType": 0,
+#         "replayOssStatus": 0,
+#         "presentations": [
+#             {
+#                 "id": "1267751453966295552",
+#                 "title": "计算机网络原理  课堂提问",
+#                 "cover": "https://thu-private-qn.yuketang.cn/slide/11789532/cover435_20241015115349.jpg?e=1729790227&token=IAM-gs8ue1pDIGwtR1CR0Zjdagg7Q2tn5G_1BqTmhmqa:CkT5RMnhXFxcVUWoxVU2xxjA6ac=",
+#                 "slidesCount": 21,
+#                 "totalSlidesCount": 59,
+#                 "doubtCount": 0,
+#                 "collectCount": 0,
+#                 "conf": "{\"show_presentation\":\"film\",\"slides\":[\"1267751453966295553\",\"1267751453974684160\",\"1267751453974684162\",\"1267751453974684164\",\"1267751453983072768\",\"1267751453983072770\",\"1267751453983072772\",\"1267751453983072774\",\"1267751453991461376\",\"1267751453991461378\",\"1267751453991461379\",\"1267751453999849984\",\"1267751453999849985\",\"1267751453999849986\",\"1267751453999849987\",\"1267751454008238592\",\"1267751454008238593\",\"1267751454008238594\",\"1267751454008238595\",\"1267751454016627200\",\"1267751454016627201\"],\"hide_slides\":[]}"
+#             }
+#         ],
+#         "user": {
+#             "identityId": "21640720",
+#             "avatar": "http://qn-sx.yuketang.cn/tougao_pic_AFzVVZcN5p9.png",
+#             "name": "董业恺",
+#             "number": "2022010426"
+#         },
+#         "activityId": "7970721",
+#         "memoContent": "",
+#         "liveViewed": false,
+#         "doubtSlides": [],
+#         "collectSlides": [],
+#         "checkIn": {
+#             "lessonId": "1267751345493205504",
+#             "identityId": "21640720",
+#             "score": 1000,
+#             "source": 5,
+#             "time": 1728964549329,
+#             "valid": 1,
+#             "problemScore": 1000,
+#             "quizScore": -1,
+#             "duration": 0,
+#             "addScore": null,
+#             "redEnvelope": 0,
+#             "correctCount": 10,
+#             "incorrectCount": 4,
+#             "unMarkCount": 0
+#         },
+#         "quizzes": [],
+#         "danmuList": [],
+#         "tougaoList": [],
+#         "toastType": 0,
+#         "problems": [
+#             {
+#                 "problemId": "1267751453974684162",
+#                 "problemType": 1,
+#                 "problemScore": 100,
+#                 "index": 3,
+#                 "cover": "https://thu-private-qn.yuketang.cn/slide/11789532/cover433_20241015115349.jpg?e=1729790227&token=IAM-gs8ue1pDIGwtR1CR0Zjdagg7Q2tn5G_1BqTmhmqa:hLi2EEWUQHgVeKyP9y4bi7neaYQ=",
+#                 "presentationId": "1267751453966295552",
+#                 "answer": [
+#                     "D"
+#                 ],
+#                 "ans_type": "",
+#                 "comment": {},
+#                 "correctAnswer": [
+#                     "D"
+#                 ],
+#                 "score": 100,
+#                 "submitTime": 1728964586371,
+#                 "scoreTime": 0,
+#                 "correct": true,
+#                 "blankStatus": [],
+#                 "anonymous": null,
+#                 "remarkDetail": {},
+#                 "teamInfo": null
+#             }
+#         ]
+#     }
+# }
+
+
+# --- --- --- Section Download PPT --- --- --- #
+# {
+#     "code": 0,
+#     "msg": "OK",
+#     "data": {
+#         "presentation": {
+#             "id": "714674183600571776",
+#             "title": "L1_课程介绍",
+#             "cover": "https://qn-st0.yuketang.cn/FudgWS2XoU3bXLxReeSBBhYTWJsX",
+#             "width": 720,
+#             "height": 540,
+#             "conf": {
+#                 "show_presentation": "all",
+#                 "slides": [
+#                     "714674183617348992"
+#                 ],
+#                 "hide_slides": []
+#             }
+#         },
+#         "slides": [
+#             {
+#                 "id": "714674183617348992",
+#                 "index": 1,
+#                 "doubtCount": 0,
+#                 "collectCount": 0,
+#                 "cover": "https://qn-st0.yuketang.cn/FudgWS2XoU3bXLxReeSBBhYTWJsX",
+#                 "problem": null,
+#                 "result": null
+#             }
+#         ]
+#     }
+# }
+
